+++ lib/src/code_generators/swagger_additions_generator.dart
@@ -35,7 +35,7 @@ class SwaggerAdditionsGenerator {
 
       final fileName = key.split('/').last;
       maps.writeln('  ...$className,');
-      imports.add("import '$fileName.dart';");
+      imports.add("import '${fileName.replaceAll('-', '_')}.dart';");
     });
 
     imports.sort();