+++ lib/src/code_generators/swagger_requests_generator.dart
@@ -143,7 +143,7 @@ $allMethodsContent
               inParameter: 'body',
               name: 'body',
               isRequired: true,
-              type: _getBodyParameterType(additionalParameter),
+              type: _getBodyParameterType(additionalParameter, options),
               ref: additionalParameter?.ref ??
                   additionalParameter?.items?.ref ??
                   ''));
@@ -187,6 +187,7 @@ $allMethodsContent
           allEnumNames: allEnumNames,
           requestType: swaggerRequest.type,
           useRequiredAttribute: options.useRequiredAttributeForHeaders,
+          options: options,
         );
 
         final hasEnums = swaggerRequest.parameters.any((parameter) =>
@@ -208,12 +209,14 @@ $allMethodsContent
             getParameterCommentsForMethod(swaggerRequest.parameters, options);
 
         final returnTypeName = getReturnTypeName(
-            swaggerRequest.responses,
-            swaggerPath.path,
-            swaggerRequest.type,
-            options.responseOverrideValueMap,
-            dynamicResponses,
-            basicTypesMap);
+          swaggerRequest.responses,
+          swaggerPath.path,
+          swaggerRequest.type,
+          options.responseOverrideValueMap,
+          dynamicResponses,
+          basicTypesMap,
+          options,
+        );
 
         final generatedMethod = getMethodContent(
             summary: swaggerRequest.summary,
@@ -237,7 +240,8 @@ $allMethodsContent
     return methods.toString();
   }
 
-  String _getBodyParameterType(RequestContent? content) {
+  String _getBodyParameterType(
+      RequestContent? content, GeneratorOptions options) {
     if (content == null) {
       return 'Object';
     }
@@ -247,7 +251,8 @@ $allMethodsContent
         return 'Object';
       }
 
-      final type = content.items!.ref.split('/').last.capitalize;
+      final type =
+          content.items!.ref.split('/').last.capitalize + options.modelPostfix;
 
       return 'List<$type>';
     }
@@ -349,6 +354,8 @@ $allMethodsContent
                 parameter.schema?.enumValues.isNotEmpty == true))
         .map((e) => e.name)
         .toList();
+
+    parameterName = validateParameterName(parameterName);
 
     final mapName = getMapName(requestPath, requestType, parameterName, ref);
 
@@ -500,8 +507,13 @@ $allMethodsContent
     }
   }
 
-  String getBodyParameter(SwaggerRequestParameter parameter, String path,
-      String requestType, List<String> allEnumNames) {
+  String getBodyParameter(
+    SwaggerRequestParameter parameter,
+    String path,
+    String requestType,
+    List<String> allEnumNames,
+    GeneratorOptions options,
+  ) {
     String parameterType;
     if (parameter.type.isNotEmpty) {
       parameterType = parameter.type;
@@ -517,13 +529,16 @@ $allMethodsContent
 
       if (allEnumNames.contains('enums.$parameterType')) {
         parameterType = 'enums.$parameterType';
+      } else {
+        parameterType += options.modelPostfix;
       }
 
       if (parameter.type == 'array') {
         parameterType = 'List<$parameterType>';
       }
     } else if (parameter.schema?.ref.isNotEmpty ?? false) {
-      parameterType = parameter.schema!.ref.split('/').last;
+      parameterType =
+          parameter.schema!.ref.split('/').last + options.modelPostfix;
     } else {
       parameterType = defaultBodyParameter;
     }
@@ -554,17 +569,25 @@ $allMethodsContent
     return "@${parameter.inParameter.capitalize}('${parameter.name}') ${parameter.isRequired ? "@required" : ""} $parameterType? ${validateParameterName(parameter.name)}";
   }
 
-  String getParameterContent(
-      {required SwaggerRequestParameter parameter,
-      required bool ignoreHeaders,
-      required String requestType,
-      required String path,
-      required List<String> allEnumNames,
-      required bool useRequiredAttribute}) {
+  String getParameterContent({
+    required SwaggerRequestParameter parameter,
+    required bool ignoreHeaders,
+    required String requestType,
+    required String path,
+    required List<String> allEnumNames,
+    required bool useRequiredAttribute,
+    required GeneratorOptions options,
+  }) {
     final parameterType = validateParameterType(parameter.name);
     switch (parameter.inParameter) {
       case 'body':
-        return getBodyParameter(parameter, path, requestType, allEnumNames);
+        return getBodyParameter(
+          parameter,
+          path,
+          requestType,
+          allEnumNames,
+          options,
+        );
       case 'formData':
         final isEnum = parameter.schema?.enumValues.isNotEmpty ?? false;
 
@@ -572,9 +595,18 @@ $allMethodsContent
       case 'header':
         final needRequiredAttribute =
             parameter.isRequired && useRequiredAttribute;
+
+        final defaultValue = options.defaultHeaderValuesMap.firstWhereOrNull(
+            (element) =>
+                element.headerName.toLowerCase() ==
+                parameter.name.toLowerCase());
+
+        final defaultValuePart =
+            defaultValue == null ? '' : ' = \'${defaultValue.defaultValue}\'';
+
         return ignoreHeaders
             ? ''
-            : "@${parameter.inParameter.capitalize}('${parameter.name}') ${needRequiredAttribute ? "required" : ""} String? ${validateParameterName(parameter.name)}";
+            : "@Header('${parameter.name}') ${needRequiredAttribute ? "required" : ""} String? ${validateParameterName(parameter.name)}$defaultValuePart";
       case 'cookie':
         return '';
       default:
@@ -619,13 +651,15 @@ abstract class $className extends ChopperService''';
     return classWithoutChopper;
   }
 
-  String getAllParametersContent(
-      {required List<SwaggerRequestParameter> listParameters,
-      required bool ignoreHeaders,
-      required String path,
-      required String requestType,
-      required List<String> allEnumNames,
-      required bool useRequiredAttribute}) {
+  String getAllParametersContent({
+    required List<SwaggerRequestParameter> listParameters,
+    required bool ignoreHeaders,
+    required String path,
+    required String requestType,
+    required List<String> allEnumNames,
+    required bool useRequiredAttribute,
+    required GeneratorOptions options,
+  }) {
     return listParameters
         .map((SwaggerRequestParameter parameter) => getParameterContent(
               parameter: parameter,
@@ -634,6 +668,7 @@ abstract class $className extends ChopperService''';
               allEnumNames: allEnumNames,
               requestType: requestType,
               useRequiredAttribute: useRequiredAttribute,
+              options: options,
             ))
         .where((String element) => element.isNotEmpty)
         .join(', ');
@@ -650,17 +685,21 @@ abstract class $className extends ChopperService''';
   String getResponseModelName(String url, String methodName) {
     final urlString = url.split('/').map((e) => e.pascalCase).join();
     final methodNamePart = methodName.pascalCase;
+    final responseType = SwaggerModelsGenerator.getValidatedClassName(
+        '$urlString$methodNamePart\$Response');
 
-    return '$urlString$methodNamePart\$Response';
+    return responseType;
   }
 
   String getReturnTypeName(
-      List<SwaggerResponse> responses,
-      String url,
-      String methodName,
-      List<ResponseOverrideValueMap> overriddenRequests,
-      List<String> dynamicResponses,
-      Map<String, String> basicTypesMap) {
+    List<SwaggerResponse> responses,
+    String url,
+    String methodName,
+    List<ResponseOverrideValueMap> overriddenRequests,
+    List<String> dynamicResponses,
+    Map<String, String> basicTypesMap,
+    GeneratorOptions options,
+  ) {
     if (overriddenRequests
             .any((ResponseOverrideValueMap element) => element.url == url) ==
         true) {
@@ -684,16 +723,21 @@ abstract class $className extends ChopperService''';
     }
 
     if (neededResponse.schema?.type.isNotEmpty ?? false) {
-      final param = neededResponse.schema?.items?.originalRef.isNotEmpty == true
+      var param = neededResponse.schema?.items?.originalRef.isNotEmpty == true
           ? neededResponse.schema?.items?.originalRef
-          : neededResponse.schema?.items?.type.isNotEmpty == true
-              ? neededResponse.schema?.items?.type
-              : neededResponse.schema?.items?.ref.split('/').lastOrNull ?? '';
-      return getParameterTypeName(neededResponse.schema?.type ?? '', param!);
+          : neededResponse.schema?.items?.type ?? '';
+
+      if (param!.isEmpty &&
+          neededResponse.schema?.items?.ref.split('/').lastOrNull != null) {
+        param = neededResponse.schema!.items!.ref.split('/').last +
+            options.modelPostfix;
+      }
+
+      return getParameterTypeName(neededResponse.schema?.type ?? '', param);
     }
 
     if (neededResponse.schema?.ref.isNotEmpty ?? false) {
-      return neededResponse.schema?.ref.split('/').lastOrNull ?? '';
+      return neededResponse.schema!.ref.split('/').last + options.modelPostfix;
     }
 
     if (neededResponse.ref.isNotEmpty) {
@@ -703,7 +747,7 @@ abstract class $className extends ChopperService''';
           dynamicResponses.contains(ref)) {
         return 'object';
       } else {
-        return ref;
+        return ref + options.modelPostfix;
       }
     }
 
@@ -714,7 +758,11 @@ abstract class $className extends ChopperService''';
     if (neededResponse.content.isNotEmpty == true &&
         neededResponse.content.isNotEmpty) {
       if (neededResponse.content.first.ref.isNotEmpty) {
-        final type = neededResponse.content.first.ref.split('/').last;
+        var type = neededResponse.content.first.ref.split('/').last;
+
+        if (!basicTypesMap.containsKey(type)) {
+          type += options.modelPostfix;
+        }
 
         if (basicTypesMap.containsKey(type.toString())) {
           return basicTypesMap[type] ?? '';
@@ -722,13 +770,15 @@ abstract class $className extends ChopperService''';
         return type;
       }
       if (neededResponse.content.first.responseType.isNotEmpty) {
-        return getParameterTypeName(
-            neededResponse.content.first.responseType,
-            neededResponse.schema?.items?.originalRef ??
-                neededResponse.content.firstOrNull?.items?.ref
-                    .split('/')
-                    .lastOrNull ??
-                '');
+        final ref = neededResponse.content.firstOrNull?.items?.ref
+            .split('/')
+            .lastOrNull;
+        if (ref?.isNotEmpty == true) {
+          return getParameterTypeName(neededResponse.content.first.responseType,
+              ref! + options.modelPostfix);
+        }
+        return getParameterTypeName(neededResponse.content.first.responseType,
+            neededResponse.schema?.items?.originalRef ?? '');
       }
     }