+++ lib/src/models/generator_options.dart
@@ -5,21 +5,24 @@ part 'generator_options.g2.dart';
 @JsonSerializable(fieldRename: FieldRename.snake, anyMap: true)
 class GeneratorOptions {
   /// Instantiate generator options.
-  GeneratorOptions(
-      {this.withBaseUrl = true,
-      this.withConverter = true,
-      this.ignoreHeaders = false,
-      this.useDefaultNullForLists = false,
-      this.buildOnlyModels = false,
-      this.defaultValuesMap = const <DefaultValueMap>[],
-      this.responseOverrideValueMap = const <ResponseOverrideValueMap>[],
-      required this.inputFolder,
-      required this.outputFolder,
-      this.enumsCaseSensitive = true,
-      this.usePathForRequestNames = false,
-      this.useRequiredAttributeForHeaders = true,
-      this.useInheritance = true,
-      this.includeIfNull});
+  GeneratorOptions({
+    this.withBaseUrl = true,
+    this.withConverter = true,
+    this.ignoreHeaders = false,
+    this.useDefaultNullForLists = false,
+    this.buildOnlyModels = false,
+    this.defaultValuesMap = const <DefaultValueMap>[],
+    this.defaultHeaderValuesMap = const <DefaultHeaderValueMap>[],
+    this.responseOverrideValueMap = const <ResponseOverrideValueMap>[],
+    required this.inputFolder,
+    required this.outputFolder,
+    this.enumsCaseSensitive = true,
+    this.usePathForRequestNames = false,
+    this.useRequiredAttributeForHeaders = true,
+    this.useInheritance = true,
+    this.includeIfNull,
+    this.modelPostfix = '',
+  });
 
   /// Build options from a JSON map.
   factory GeneratorOptions.fromJson(Map<String, dynamic> json) =>
@@ -61,8 +64,14 @@ class GeneratorOptions {
   @JsonKey(defaultValue: false)
   final bool buildOnlyModels;
 
+  @JsonKey(defaultValue: '')
+  final String modelPostfix;
+
   @JsonKey(defaultValue: <DefaultValueMap>[])
   final List<DefaultValueMap> defaultValuesMap;
+
+  @JsonKey(defaultValue: <DefaultHeaderValueMap>[])
+  final List<DefaultHeaderValueMap> defaultHeaderValuesMap;
 
   @JsonKey(defaultValue: <ResponseOverrideValueMap>[])
   final List<ResponseOverrideValueMap> responseOverrideValueMap;
@@ -125,4 +134,20 @@ class ResponseOverrideValueMap {
 
   /// Convert this default value map instance to JSON.
   Map<String, dynamic> toJson() => _$ResponseOverrideValueMapToJson(this);
+}
+
+@JsonSerializable(fieldRename: FieldRename.snake)
+class DefaultHeaderValueMap {
+  DefaultHeaderValueMap({required this.headerName, required this.defaultValue});
+
+  @JsonKey(defaultValue: '')
+  final String headerName;
+
+  @JsonKey(defaultValue: '')
+  final String defaultValue;
+
+  Map<String, dynamic> toJson() => _$DefaultHeaderValueMapToJson(this);
+
+  factory DefaultHeaderValueMap.fromJson(Map<String, dynamic> json) =>
+      _$DefaultHeaderValueMapFromJson(json);
 }